Move controls handlers into separate files for readability
