--- handlers/controls.go
@@ -1,95 +1,0 @@
-package handlers
-
-import (
-	"net/http"
-
-	"github.com/golang/glog"
-)
-
-type Nexter interface {
-	Next() error
-}
-
-func NextHandler(c Nexter) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		err := c.Next()
-		if err != nil {
-			glog.Errorln(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-}
-
-type Previouser interface {
-	Previous() error
-}
-
-func PreviousHandler(c Previouser) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		err := c.Previous()
-		if err != nil {
-			glog.Errorln(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-}
-
-type Player interface {
-	Play(pos int) error
-}
-
-func PlayHandler(c Player) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		err := c.Play(-1)
-		if err != nil {
-			glog.Errorln(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-}
-
-type Pauser interface {
-	Pause(pause bool) error
-}
-
-func PauseHandler(c Pauser) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		err := c.Pause(true)
-		if err != nil {
-			glog.Errorln(err)
-			w.WriteHeader(http.StatusInternalServerError)
-			return
-		}
-		w.WriteHeader(http.StatusOK)
-	})
-}
-
-type Randomer interface {
-	Random(random bool) error
-}
-
-func random(c Randomer, on bool, w http.ResponseWriter) {
-	err := c.Random(on)
-	if err != nil {
-		glog.Errorln(err)
-		w.WriteHeader(http.StatusInternalServerError)
-		return
-	}
-	w.WriteHeader(http.StatusOK)
-}
-func RandomOnHandler(c Randomer) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		random(c, true, w)
-	})
-}
-func RandomOffHandler(c Randomer) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		random(c, false, w)
-	})
-}
