+++ handlers/random.go
@@ -1,0 +1,31 @@
+package handlers
+
+import (
+	"net/http"
+
+	"github.com/golang/glog"
+)
+
+type Randomer interface {
+	Random(random bool) error
+}
+
+func random(c Randomer, on bool, w http.ResponseWriter) {
+	err := c.Random(on)
+	if err != nil {
+		glog.Errorln(err)
+		w.WriteHeader(http.StatusInternalServerError)
+		return
+	}
+	w.WriteHeader(http.StatusOK)
+}
+func RandomOnHandler(c Randomer) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		random(c, true, w)
+	})
+}
+func RandomOffHandler(c Randomer) http.Handler {
+	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		random(c, false, w)
+	})
+}
